requirements_full.txt
# CourseSumm v2 - Full Requirements (All Phases)

# Phase 1 - Core Pipeline
openai>=1.0.0
anthropic>=0.7.0
openai-whisper>=20231117
torch>=2.0.0
python-docx>=0.8.11
pyyaml>=6.0
tqdm>=4.66.0
ffmpeg-python>=0.2.0

# Phase 2 - Professional Formatting
Pillow>=10.0.0

# Phase 4 - Gradio UI
gradio>=4.0.0

# Optional: Google Drive integration
# google-auth>=2.23.0
# google-auth-oauthlib>=1.1.0
# google-auth-httplib2>=0.1.1
# google-api-python-client>=2.100.0
